+++ src/main/java/org/aarboard/nextcloud/api/filesharing/SingleRemoteShareXMLAnswer.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright (C) 2026 a.schild
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package org.aarboard.nextcloud.api.filesharing;
+
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlRootElement;
+import org.aarboard.nextcloud.api.utils.XMLAnswer;
+
+/**
+ * Answer of the "get single federated/remote share" call.
+ *
+ * @author a.schild
+ */
+@XmlRootElement(name = "ocs")
+public class SingleRemoteShareXMLAnswer extends XMLAnswer
+{
+    @XmlElement(name = "data")
+    private RemoteShare remoteShare = null;
+
+    public RemoteShare getRemoteShare() {
+        return remoteShare;
+    }
+}